Separate out and standardize models, agents, and prompts
