+++ backend/service/agents/__init__.py
@@ -1,0 +1,4 @@
+from service.agents.comm_agent import CommunicationAgent
+from service.agents.memory_agent import MemoryAgent
+from service.agents.voice_agent import VoiceCommunicationAgent
+from service.agents.voice_memory_agent import VoiceMemoryAgent
