+++ backend/service/prompts/__init__.py
@@ -1,0 +1,2 @@
+from service.prompts.comm_prompts import *
+from service.prompts.memory_prompts import *
